Make claims plugin also work in non-vite mode
